assr/cyberagents/requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
google-api-core>=2.0.0
colorama>=0.4.0
